requirements/dev.txt
-r ./base.txt

django-environ==0.4.5
Faker==2.0.4
django-filter==2.2.0
djangorestframework==3.11.0
Markdown==3.1.1
pylint==2.4.4
redis==3.5.3
djangorestframework-simplejwt==4.4.0
djecrety==1.0.11
django-cors-headers==3.2.0
drf-yasg==1.17.1
